freshwater_fish_disease/requirements.txt
torch>=2.1
torchvision>=0.16
torchaudio>=2.1
torchmetrics>=1.3
tqdm>=4.66
albumentations>=1.4
opencv-python>=4.8
scikit-learn>=1.5
matplotlib>=3.8
pillow>=10.2
pandas>=2.2
pyyaml>=6.0
transformers>=4.44
tokenizers>=0.15
fastapi>=0.112
uvicorn>=0.30
gradio>=4.42
